cleanup and rename finished / reproduced / completed
