+++ paraffin/cli.py
@@ -7,6 +7,21 @@ import json
 
 import typer
 
+import threading
+import socket
+import os
+import time
+import subprocess
+import json
+from typing import Optional
+from paraffin.db.app import (
+    register_worker,
+    close_worker,
+    get_job,
+    update_job,
+    StageStatus,
+)
+
 app = typer.Typer()
 
 
@@ -37,7 +52,7 @@ def commit():
             worker_id=worker_id,
             experiment=None,
             stage_name=None,
-            status=[StageStatus.REPRODUCED],
+            status=[StageStatus.FINISHED],
         )
         if res is None:
             break
@@ -49,7 +64,7 @@ def commit():
         update_job(
             db_url=db,
             stage_id=job.stage_id,
-            status=StageStatus.FINISHED,
+            status=StageStatus.COMPLETED,
         )
 
     print("No job found.")
@@ -57,63 +72,7 @@ def commit():
 
 
 
-import threading
-import socket
-import os
-import time
-import subprocess
-import json
-from typing import Optional
-from paraffin.db.app import (
-    register_worker,
-    close_worker,
-    get_job,
-    update_job,
-    StageStatus,
-)
 
-def run_worker(name: str, db: str):
-    worker_id = register_worker(
-        name=name,
-        machine=socket.gethostname(),
-        db_url=db,
-        cwd=os.getcwd(),
-        pid=os.getpid(),
-    )
-
-    try:
-        while True:
-            res = get_job(
-                db_url=db,
-                queues=None,
-                worker_id=worker_id,
-                experiment=None,
-                stage_name=None,
-                status=[StageStatus.QUEUED],
-            )
-            if res is None:
-                break
-
-            stage, job = res
-            cmd = json.loads(stage.cmd)
-            print(f"({worker_id}) Running command: {cmd}")
-            try:
-                subprocess.check_call(cmd, shell=True)
-                update_job(
-                    db_url=db,
-                    stage_id=job.stage_id,
-                    status=StageStatus.REPRODUCED,
-                )
-            except subprocess.CalledProcessError:
-                print(f"({worker_id}) Command failed: {cmd}")
-                update_job(
-                    db_url=db,
-                    stage_id=job.stage_id,
-                    status=StageStatus.FAILED,
-                )
-    finally:
-        close_worker(id=worker_id, db_url=db)
-        print(f"({worker_id}) Worker closed.")
 
 @app.command()
 def worker(
@@ -144,6 +103,7 @@ def worker(
     jobs: int = typer.Option(1, "--jobs", "-j", help="Number of jobs to run."),
 ):
     """Start a paraffin worker to process the queued DVC stages."""
+    from paraffin.worker import run_worker
 
     threads = []
     for _ in range(jobs):
@@ -158,49 +118,6 @@ def worker(
     for t in threads:
         t.join()
     print("All workers done.")
-
-    # from paraffin.db.app import get_job, register_worker, update_job, close_worker
-    # from paraffin.dvc import StageStatus
-
-    # worker_id = register_worker(
-    #     name=name,
-    #     machine=socket.gethostname(),
-    #     db_url=db,
-    #     cwd=os.getcwd(),
-    #     pid=os.getpid(),
-    # )
-    # while True:
-    #     res = get_job(
-    #         db_url=db,
-    #         queues=None,
-    #         worker_id=worker_id,
-    #         experiment=None,
-    #         stage_name=None,
-    #         status=[StageStatus.QUEUED],
-    #     )
-    #     if res is None:
-    #         break
-        
-    #     stage, job = res
-    #     cmd = json.loads(stage.cmd)
-    #     print(f"({worker_id}) Running command: {cmd}")
-    #     try:
-    #         subprocess.check_call(cmd, shell=True)
-    #         update_job(
-    #             db_url=db,
-    #             stage_id=job.stage_id,
-    #             status=StageStatus.REPRODUCED,
-    #         )
-    #     except subprocess.CalledProcessError as e:
-    #         print(f"({worker_id}) Command failed: {cmd}")
-    #         update_job(
-    #             db_url=db,
-    #             stage_id=job.stage_id,
-    #             status=StageStatus.FAILED,
-    #         )
-
-    # print("No job found.")
-    # close_worker(id=worker_id, db_url=db)
 
   
 @app.command()
